+++ src/unique.rs
@@ -3,10 +3,10 @@ use crate::{component::Unique, memory_usage::StorageMemoryUsage, storage::Storag
 /// Unique storage.
 pub struct UniqueStorage<T: Unique> {
     pub(crate) value: T,
-    pub(crate) insert: u32,
-    pub(crate) modification: u32,
-    pub(crate) last_insert: u32,
-    pub(crate) last_modification: u32,
+    pub(crate) insert: u64,
+    pub(crate) modification: u64,
+    pub(crate) last_insert: u64,
+    pub(crate) last_modification: u64,
 }
 
 impl<T: Unique> Storage for UniqueStorage<T> {
@@ -24,7 +24,7 @@ impl<T: Unique> Storage for UniqueStorage<T> {
 }
 
 impl<T: Unique> UniqueStorage<T> {
-    pub(crate) fn new(value: T, current: u32) -> Self {
+    pub(crate) fn new(value: T, current: u64) -> Self {
         UniqueStorage {
             value,
             insert: current,
